+++ src/lightcone_particle_io.h
@@ -93,9 +93,9 @@ inline static struct lightcone_io_props lightcone_io_make_output_field(
 struct lightcone_gas_data {
   long long id;
   double x[3];
-  double mass;
-  double h;
-  double a;
+  float mass;
+  float h;
+  float a;
 };
 
 int lightcone_store_gas(const struct engine *e,
@@ -110,8 +110,8 @@ int lightcone_store_gas(const struct engine *e,
 struct lightcone_dark_matter_data {
   long long id;
   double x[3];
-  double mass;
-  double a;
+  float mass;
+  float a;
 };
 
 int lightcone_store_dark_matter(const struct engine *e,
@@ -125,8 +125,8 @@ int lightcone_store_dark_matter(const struct engine *e,
 struct lightcone_stars_data {
   long long id;
   double x[3];
-  double mass;
-  double a;
+  float mass;
+  float a;
 };
 
 int lightcone_store_stars(const struct engine *e,
@@ -141,8 +141,8 @@ int lightcone_store_stars(const struct engine *e,
 struct lightcone_black_hole_data {
   long long id;
   double x[3];
-  double mass;
-  double a;
+  float mass;
+  float a;
 };
 
 int lightcone_store_black_hole(const struct engine *e,
@@ -157,8 +157,8 @@ int lightcone_store_black_hole(const struct engine *e,
 struct lightcone_neutrino_data {
   long long id;
   double x[3];
-  double mass;
-  double a;
+  float mass;
+  float a;
 };
 
 int lightcone_store_neutrino(const struct engine *e,
